- add average low and high temperatures for all cities in the combo box - display ? (NaN) for heating and cooling energy values when none of the combo box cities are selected - automatically set the selected city to none when the value of latitude spinner is modified - let the Auto check box control the value of outside temperature only

+++ src/main/java/org/concord/energy3d/shapes/Heliodon.java
@@ -65,7 +65,7 @@ import com.ardor3d.ui.text.BMText.AutoScale;
 import com.ardor3d.util.geom.BufferUtils;
 
 public class Heliodon {
-	public static final double DEFAULT_LATITUDE = 42;
+	public static final int DEFAULT_LATITUDE = 42;
 	private static final int BASE_DIVISIONS = 72;
 	private static final int DECLINATION_DIVISIONS = 12;
 	private static final int HOUR_DIVISIONS = 96;
@@ -236,7 +236,6 @@ public class Heliodon {
 		logicalLayer.registerTrigger(new InputTrigger(new MouseButtonPressedCondition(MouseButton.LEFT), new TriggerAction() {
 			@Override
 			public void perform(final Canvas source, final TwoInputStates inputStates, final double tpf) {
-//				lock = true;
 				final int x = inputStates.getCurrent().getMouseState().getX();
 				final int y = inputStates.getCurrent().getMouseState().getY();
 				final Ray3 pickRay = SceneManager.getInstance().getCanvas().getCanvasRenderer().getCamera().getPickRay(new Vector2(x, y), false, null);
@@ -261,7 +260,6 @@ public class Heliodon {
 				if (!forceSunRegionOn)
 					sunRegion.getSceneHints().setCullHint(CullHint.Always);
 				SceneManager.getInstance().setMouseControlEnabled(true);
-//				lock = false;
 			}
 		}));
